requirements.txt
python-dotenv
langchain
langchain_groq
langchain_community
flask
streamlit
plotly
networkx